requirements.txt
flask
Flask-SQLAlchemy
Flask-Testing
Flask-API